requirements.txt
numpy==1.17.4  # numerical operations
scipy==1.4.0  # temporary mat file saving for MCR binary inputs
pandas==0.24.2  # tabular data processing
matplotlib==3.0.3  # plotting
json_tricks==3.12.1  # saving json files with classes and numpy arrays
eyeD3==0.8.11  # reading metadata embedded in the audio recordings
python-Levenshtein==0.12.0  # semiotic structure labeling similarity
networkx==1.11  # semiotic structure labeling clique computation
lxml==4.4.2  # musicxml conversion
musicbrainzngs==0.6  # metadata crawling from musicbrainz
essentia==2.1b5;platform_system=='Linux'  # music signal processing